tgbot/requirements.txt
aiogram==2.13
aioredis==1.3.1
APScheduler==3.7.0
asyncpg==0.23.0
feedparser==6.0.2
googlesearch-python==1.0.1